Refactor Meta file handling
